move tree filling fn out to models
