+++ desci-models/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./ResearchObject";
 export * from "./constants";
 export * from "./transformers/index";
+export * from "./treeTools";
